allow the use of wakeup() in interrupts
inferno native driver processes waiting on an interrupt use
busy sleep unlike 9front drivers. The 9front drivers wakeup()
any processes waiting on the interrupt completion.

The wakeup() uses lock()/unlock(). unlock() calls sched()
to kick off other Ready processes. But, this cannot be done
in an interrupt handler. The interrupt needs to complete first.
Hence, the islo() check to ensure that we are not in an interrupt
handler.